Hice las cosa de usuario

+++ src/api/users/user.service.js
@@ -1,35 +1,61 @@
+const { hashPassword } = require('../../../shared/utils/util');
 const userRepository = require('./user.repository');
-const { hashPassword} = require('../../../shared/utils/util');
-const { get } = require('../auth/auth.routes');
 
+const getPaginatedUsers = async (page= 1, limit =10) => {
+    const offset = (page - 1) * limit;
+
+    const {count, rows} = await userRepository.getPaginatedUsers(limit,offset);
+
+    return {
+        totalItems: count,
+        totalPages: Math.ceil(count / limit),
+        currentPage: page,
+        users: rows
+    };
+};
+
+const getUserById = async (id) => {
+    const user = await userRepository.findUserById(id);
+    if (!user) throw new Error("User not found");
+    return user;
+};
 
 const getAllUsers = async () => {
     return await userRepository.getAllUsers();
 };
 
-const getUserById = async (id) => {
-    return await userRepository.getUserById(id);
-};
-
 const createUser = async (userData) => {
-    const hashPassword = await hashPassword(userData.password);
-    return await userRepository.createUser({...userData, password:hashedPassword});
+    const hashedPassword = await hashPassword(userData.password);
+    return await userRepository.createUser({
+        ...userData, 
+        password: hashedPassword
+    });
 }
 
 const updateUser = async (id, userData) => {
-    if (userData.password){
+    const user = await userRepository.findUserById(id);
+    if (!user) throw new Error("User not found");
+
+    if (userData.password) {
         userData.password = await hashPassword(userData.password);
     }
+
     await userRepository.updateUser(id, userData);
+
     return await userRepository.findUserById(id);
 };
 
 const deleteUser = async (id) => {
-    return await userRepository.deleteUser(id);
+    const user = await userRepository.findUserById(id);
+    if (!user) throw new Error("User not found");
+
+    await userRepository.deleteUser(id);
+    return user;
 }
 
 module.exports = {
     getAllUsers,
+    getPaginatedUsers,
     getUserById,
     createUser,
     updateUser,
